backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
youtube-transcript-api==0.6.2
openai>=1.16.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.2 